Split the contents dom.j into 4 files: dom.js, p5.Element.js, p5.File.js, p5.MediaElements.js
